\name{HealthTest}
\alias{HealthTest}
\title{Some title}
\usage{
  HealthTest(strategy, pars, dates = NULL,
    portfolio = Test(strategy, pars, dates = dates, details = rtype == "trades", returns = rtype == "periods"),
    R = Returns(portfolio, type = rtype[1], reduce = T),
    rtype = c("periods", "trades"),
    test = c("runs", "accuracy", "mean"), w = 0.6,
    mu = c(0, "Random"))
}
\description{
  Some title
}

