\name{plot.xts.custom}
\alias{plot.xts.custom}
\title{Some title}
\usage{
  plot.xts.custom(x, y = NULL, screens = "auto",
    layout.screens = "auto", ...,
    yax.loc = c("none", "out", "in", "flip", "left", "right", "top"),
    auto.grid = TRUE, major.ticks = "auto",
    minor.ticks = TRUE, major.format = TRUE,
    bar.col.up = "white", bar.col.dn = "red",
    candle.col = "black", xy.labels = FALSE,
    xy.lines = NULL, ylim = "auto", panel = default.panel,
    auto.legend = FALSE, legend.names = colnames(x),
    legend.loc = "topleft", legend.pars = NULL, events,
    blocks, nc, nr)
}
\description{
  Some title
}

